+++ src/image/parameters.rs
@@ -97,7 +97,7 @@ pub fn process_args(args: Vec<String>) -> Option<(Scene, Parameters)> {
     match scene_index {
         Some(scene_index) => {
             let aspect_ratio =
-                width.unwrap_or(WIDTH_DEFAULT) as f64 / height.unwrap_or(HEIGHT_DEFAULT) as f64;
+                width.unwrap_or(WIDTH_DEFAULT) as f32 / height.unwrap_or(HEIGHT_DEFAULT) as f32;
             let scene = get_scene(&args, scene_index, aspect_ratio);
 
             let parameters = Parameters::new(samples, width, height, filename);
@@ -218,7 +218,7 @@ fn get_info(args: &Vec<String>, index: usize) {
     }
 }
 
-fn get_scene(args: &Vec<String>, index: usize, aspect_ratio: f64) -> Scene {
+fn get_scene(args: &Vec<String>, index: usize, aspect_ratio: f32) -> Scene {
     match args.get(index) {
         None => {
             println!("Please specify a value for scene!");
